requirements.txt
fastapi==0.124.4
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.38.0
sqlmodel==0.0.27